Convert the upper house

+++ EU4toV2/Source/V2World/V2Country.h
@@ -111,9 +111,9 @@ class V2Country
 		double							researchPoints;
 		string							techSchool;
 		string							government;
-		double							upperHouseReactionary;
-		double							upperHouseConservative;
-		double							upperHouseLiberal;
+		int								upperHouseReactionary;
+		int								upperHouseConservative;
+		int								upperHouseLiberal;
 		vector< pair<int, int> >	reactionaryIssues;
 		vector< pair<int, int> >	conservativeIssues;
 		vector< pair<int, int> >	liberalIssues;
